requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
telebot==0.0.5
urllib3==2.0.2
wget==3.2
